backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.25.0
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.2.1
dnspython==2.8.0
httpx==0.28.1
